backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Devin AI <158243242+devin-ai-integration[bot]@users.noreply.github.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.115.6"}
psycopg = {extras = ["binary"], version = "^3.2.4"}
pandas = "^2.2.3"
numpy = ">=1.24,<2.2"
scikit-learn = "^1.6.1"
torch = {version = "^2.1.0", source = "pytorch", extras = ["cpu"]}
sentence-transformers = {version = "^3.3.1", extras = ["cpu"]}
faiss-cpu = "^1.9.0.post1"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.7.1"
psutil = "^5.9.8"  # For memory monitoring
orjson = "^3.9.10"  # Faster JSON processing

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
